backend/requirements.txt
# Backend dependencies
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary==2.9.9
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Authentication dependencies
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# RAG dependencies (local embeddings, no OpenAI)
langchain>=0.0.267
sentence-transformers>=2.2.2
pyarrow>=12.0.0
spacy>=3.7.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Vector databases (open source alternatives to ChromaDB)
qdrant-client>=1.7.0
weaviate-client>=3.25.0
faiss-cpu>=1.7.4  # or faiss-gpu for GPU support

# Graph database (JanusGraph via Gremlin Server)
# JanusGraph 1.0.x uses TinkerPop 3.6.x; align driver accordingly
gremlinpython==3.6.6

# LLM Integration
ollama>=0.1.0
requests>=2.31.0
httpx>=0.24.1

# Async support
asyncio-mqtt>=0.13.0
aioredis>=2.0.1

# System monitoring
psutil>=5.9.0
prometheus-client>=0.17.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.1